% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_eval.R
\name{add_angle_at_point_type}
\alias{add_angle_at_point_type}
\title{Add the degree angle between the middle index of a specific point type and the maximum XY points to the left and right of the given middle index
Uses Law of Cosines to determine angle from a given point given a set of 3 points that can form a triangle the rest of the cross section points in each hy_id, cs_id}
\usage{
add_angle_at_point_type(
  cross_section_pts,
  angle_at = "bottom",
  default_col_name = TRUE
)
}
\arguments{
\item{cross_section_pts}{cross section points dataframe with a "point_type" column and "hy_id", "cs_id" columns}

\item{angle_at}{character, which point type to get the degree angle for. Must be one of  "left_bank", "bottom", "right_bank", or "channel". Default is "bottom"}

\item{default_col_name}{logical, whether the output column should be named "angle_at" or
if the new column should take the "point_type" string and use. Default is TRUE and adds a column named "angle_at"}
}
\value{
dataframe, the cross_section_pts dataframe with an added "angle_at" column
}
\description{
Add the degree angle between the middle index of a specific point type and the maximum XY points to the left and right of the given middle index
Uses Law of Cosines to determine angle from a given point given a set of 3 points that can form a triangle the rest of the cross section points in each hy_id, cs_id
}
