% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_attribute_based_extension_distances}
\alias{add_attribute_based_extension_distances}
\title{Add an extension_distance column based off valid_banks and has_relief attributes}
\usage{
add_attribute_based_extension_distances(
  transects,
  scale = 0.5,
  length_col = NULL
)
}
\arguments{
\item{transects}{dataframe, tibble or sf dataframe with length_col,  "valid_banks", and "has_relief" columns}

\item{scale}{numeric, percentage of current transect line length to extend transects in transects_to_extend by. Default is 0.5 (50\% of the transect length)}

\item{length_col}{character, name of the column with the numeric cross section length}
}
\value{
dataframe, tibble or sf dataframe
}
\description{
Add an extension_distance column based off valid_banks and has_relief attributes
}
