% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_bank_attributes.R
\name{add_bank_attributes}
\alias{add_bank_attributes}
\title{Adds attributes about the banks of each cross section in a dataframe of cross section points
Function adds "bottom", "left_bank", "right_bank" columns that are
the Z values of the "lowest" bottom point, and the "highest" left and right bank Z values, respectively. If there are
And also a "valid_banks" column is added that is TRUE if the hy_id/cs_id set of cross section point has at least 1 bottom point with
at least 1 left bank point AND 1 right bank point that are above the lowest "bottom" point.}
\usage{
add_bank_attributes(classified_pts)
}
\arguments{
\item{classified_pts}{sf or dataframe of points with "hy_id", "cs_id", and "point_type" columns. Output of hydrofabric3D::classify_pts()}
}
\value{
sf or dataframe with added "bottom", "left_bank", "right_bank", and "valid_banks" columns
}
\description{
Adds attributes about the banks of each cross section in a dataframe of cross section points
Function adds "bottom", "left_bank", "right_bank" columns that are
the Z values of the "lowest" bottom point, and the "highest" left and right bank Z values, respectively. If there are
And also a "valid_banks" column is added that is TRUE if the hy_id/cs_id set of cross section point has at least 1 bottom point with
at least 1 left bank point AND 1 right bank point that are above the lowest "bottom" point.
}
