% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{add_braid_ids}
\alias{add_braid_ids}
\title{Find braids and add to a dataframe/sf dataframe
Adds a 'braid_id' and 'is_multibraid' columns to an sf dataframe containing a crosswalk_id and sf linestring geometires}
\usage{
add_braid_ids(network, crosswalk_id = NULL, verbose = FALSE)
}
\arguments{
\item{network}{The network object representing the river network.}

\item{crosswalk_id}{unique ID column name}

\item{verbose}{Logical indicating whether to display verbose messages during the braid detection process.}
}
\value{
dataframe or sf dataframe with added braid_id
}
\description{
Find braids and add to a dataframe/sf dataframe
Adds a 'braid_id' and 'is_multibraid' columns to an sf dataframe containing a crosswalk_id and sf linestring geometires
}
