% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_area.R
\name{add_cs_area2}
\alias{add_cs_area2}
\title{Adds a cs_area column to a set of cross section points (v2)}
\usage{
add_cs_area2(cs_pts, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts}{dataframe or sf dataframe of CS points with crosswalk_id, cs_id, Z, and relative distance columns}

\item{crosswalk_id}{character, unique ID column name}
}
\value{
cs_pts dataframe with added numeric 'cs_area' column
}
\description{
Adds a cs_area column to a set of cross section points (v2)
}
