% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahg_estimates.R
\name{add_cs_bathymetry}
\alias{add_cs_bathymetry}
\title{Given provide inchannel widths and depths to a set of cross section points and derive estimated shapes}
\usage{
add_cs_bathymetry(cross_section_pts = NULL)
}
\arguments{
\item{cross_section_pts}{dataframe or sf dataframe. Default is NULL}
}
\value{
dataframe or sf dataframe with AHG estimated points injected into the input cross section points
}
\description{
Still in early development phases
}
