% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_cs_id_sequence}
\alias{add_cs_id_sequence}
\title{Add a 1:number of cross sections 'cs_id' for each crosswalk_id by cs_measure}
\usage{
add_cs_id_sequence(x, crosswalk_id = NULL)
}
\arguments{
\item{x}{dataframe, sf dataframe or tibble}

\item{crosswalk_id}{character, unique ID column}
}
\value{
dataframe, sf dataframe or tibble with an added 'cs_id' column
}
\description{
Add a 1:number of cross sections 'cs_id' for each crosswalk_id by cs_measure
}
