% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_default_bank_attributes}
\alias{add_default_bank_attributes}
\title{Add "bottom", "left_bank", "right_bank", and "valid_banks" column defaults to a dataframe
Internal helper function for get_bank_attributes()}
\usage{
add_default_bank_attributes(df)
}
\arguments{
\item{df}{dataframe, tibble, or sf dataframe}
}
\value{
dataframe, tibble, or sf dataframe
}
\description{
Add "bottom", "left_bank", "right_bank", and "valid_banks" column defaults to a dataframe
Internal helper function for get_bank_attributes()
}
