% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_hydrofabric_id}
\alias{add_hydrofabric_id}
\title{Add a unique 'hydrofabric_id` to each row of a dataframe
Internal conveniance function for when a dataframe / flowlines network does NOT have a specified ID column}
\usage{
add_hydrofabric_id(df)
}
\arguments{
\item{df}{sf dataframe, tibble, or dataframe}
}
\value{
dataframe, sf dataframe, or tibble
}
\description{
Add a unique 'hydrofabric_id` to each row of a dataframe
Internal conveniance function for when a dataframe / flowlines network does NOT have a specified ID column
}
