% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_to_polygon_extender.R
\name{add_intersects_ids}
\alias{add_intersects_ids}
\title{Add an ID column from 'y' if it intersects with 'x'}
\usage{
add_intersects_ids(x, y, id_col)
}
\arguments{
\item{x}{sf dataframe}

\item{y}{sf dataframe}

\item{id_col}{character, unique ID column name in 'y'}
}
\value{
sf dataframe of x with id_col added if it intersects with y
}
\description{
Add an ID column from 'y' if it intersects with 'x'
}
