% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_is_complete_cs_flag}
\alias{add_is_complete_cs_flag}
\title{Add an is_complete_cs flag to cross sections points
Any cross section points that has does NOT have ANY NA Z (depth) values is flagged as is_complete_cs = TRUE}
\usage{
add_is_complete_cs_flag(cs_pts, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts}{cs points dataframe, tibble, or sf dataframe}

\item{crosswalk_id}{unique ID for flowline}
}
\value{
cross_section_pts dataframe / tibble / sf dataframe with cross section points with is_complete_cs flag added
}
\description{
Add an is_complete_cs flag to cross sections points
Any cross section points that has does NOT have ANY NA Z (depth) values is flagged as is_complete_cs = TRUE
}
