% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{add_is_missing_depth_flag}
\alias{add_is_missing_depth_flag}
\title{Add an is_missing_depth flag to cross sections points
Any cross section points that has missing Z (depth = NA) values is flagged as is_missing_depth = TRUE}
\usage{
add_is_missing_depth_flag(cs_pts)
}
\arguments{
\item{cs_pts}{cs points dataframe, tibble, or sf dataframe}
}
\value{
cross_section_pts dataframe / tibble / sf dataframe with cross section points missing depths flag added
}
\description{
Add an is_missing_depth flag to cross sections points
Any cross section points that has missing Z (depth = NA) values is flagged as is_missing_depth = TRUE
}
