% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{add_is_outlet_flag}
\alias{add_is_outlet_flag}
\title{Adds a logical 'is_outlet' flag to a set of transects identifying the most downstream transect}
\usage{
add_is_outlet_flag(x, crosswalk_id = NULL)
}
\arguments{
\item{x}{sf dataframe linestrings}

\item{crosswalk_id}{character}
}
\value{
sf dataframe of transects with added is_outlet logical column
}
\description{
Adds a logical 'is_outlet' flag to a set of transects identifying the most downstream transect
}
