% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_length_col}
\alias{add_length_col}
\title{Add a length column to a sf geometry dataframe}
\usage{
add_length_col(x, length_col = NULL, add_unit_to_col = FALSE)
}
\arguments{
\item{x}{sf dataframe}

\item{length_col}{character, name to use for length column. Default is NULL which will use "geom_length" as the length column name}

\item{add_unit_to_col}{logical, whether to try and extract units from the geometry and append these to the column name. Default is FALSE}
}
\value{
sf dataframe
}
\description{
Add a length column to a sf geometry dataframe
}
