% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_eval.R
\name{add_middle_index_by_point_type}
\alias{add_middle_index_by_point_type}
\title{Add a "middle_index" column denoting the middle index of a specific point_type
The middle index is relative to the rest of the cross section points in each hy_id, cs_id}
\usage{
add_middle_index_by_point_type(
  cross_section_pts,
  point_type = "bottom",
  default_col_name = TRUE
)
}
\arguments{
\item{cross_section_pts}{cross section points dataframe with a "point_type" column and "hy_id", "cs_id" columns}

\item{point_type}{character, which point type to get the middle index for. Must be one of  "left_bank", "bottom", "right_bank", or "channel". Default is "bottom"}

\item{default_col_name}{logical, whether the output column should be named "middle_index" or if
the new column should take the point_type string and use that in the column name (i.e. "left_bank_middle_index" instead of "middle_index").
Default is TRUE and adds a column named "middle_index"}
}
\value{
dataframe of the input cross_section_pts with an added middle index column
}
\description{
Add a "middle_index" column denoting the middle index of a specific point_type
The middle index is relative to the rest of the cross section points in each hy_id, cs_id
}
