% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_improvements.R
\name{add_needs_rectification}
\alias{add_needs_rectification}
\title{Add a "needs_rectification" column to a sf/dataframe}
\usage{
add_needs_rectification(transects)
}
\arguments{
\item{transects}{sf linestring with "valid_banks" and "has_relief" logical columns}
}
\value{
logical, TRUE if there are transects without valid banks or relief
}
\description{
Add a "needs_rectification" column to a sf/dataframe
}
