% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_bank_attributes.R
\name{add_point_type_counts}
\alias{add_point_type_counts}
\title{Add the count of each point type as a column to a dataframe of section points}
\usage{
add_point_type_counts(classified_pts, crosswalk_id = NULL)
}
\arguments{
\item{classified_pts}{dataframe or sf dataframe, cross section points with a "hy_id", and "cs_id" columns as well as a 'point_type' column containing the values: "bottom", "left_bank", "right_bank", and "channel"}

\item{crosswalk_id}{character, ID column}
}
\value{
dataframe or sf dataframe with "<point_type>_count" columns added
}
\description{
add_point_type_counts() will add columns to the input dataframe with the counts of every point_type for each hy_id/cs_id in the input dataframe of classified cross section points (output of classify_pts())
}
