% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{add_points_per_cs}
\alias{add_points_per_cs}
\title{Add a points per cross section column to an sf dataframe of linestrings given a DEM and min points value}
\usage{
add_points_per_cs(
  cs,
  points_per_cs = NULL,
  min_pts_per_cs = 10,
  dem = default_dem
)
}
\arguments{
\item{cs}{An sf dataframe representing cross-sections (linestrings). With a required cs_lengthm column (length of cross section in meters)}

\item{points_per_cs}{numeric, number of points per cross section. Default is NULL}

\item{min_pts_per_cs}{An optional minimum points value per cross section. If not provided,}

\item{dem}{A SpatRaster object representing the Digital Elevation Model (DEM) or a character string referencing a remote resource.
the function calculates it based on the length of cross-sections and the resolution of the DEM.}
}
\value{
An updated sf dataframe with the 'points_per_cs' column added.
}
\description{
This function calculates and adds a column called 'points_per_cs' to an sf dataframe
representing cross-sections (linestrings) based on a provided DEM and a minimum points
value per cross section.
}
