% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankful_width.R
\name{add_powerlaw_bankful_width}
\alias{add_powerlaw_bankful_width}
\title{Add powerlaw_bankful_width column}
\usage{
add_powerlaw_bankful_width(df, total_drainage_area_sqkm_col, min_bf_width)
}
\arguments{
\item{df}{dataframe}

\item{total_drainage_area_sqkm_col}{character, column with the total downstrream drainage area in square kilometers (numeric column)}

\item{min_bf_width}{numeric, minimum bankful width value}
}
\value{
character, column with the total downstrream drainage area in square kilometers (numeric column)
}
\description{
Add powerlaw_bankful_width column
}
