% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_relief.R
\name{add_relief}
\alias{add_relief}
\title{Add relief attributes to a dataframe of cross sections points
Given a set of cross section points (derived from hydrofabric3D::cross_section_pts() and hydrofabric3D::classify_points()) add a "has_relief" logical
value to data. The "has_relief" value is indicating whether a cross section "has relief".
Relief is determined by checking each set of cross section points have a left OR right bank that
has a depth difference from the bottom that isgreater than or equal to a percentage of the cross section length (e.g. Assuming a 'pct_of_length_for_relief' of 0.01 (1\%) of a 100m cross section would have a relief depth threshold of 1m)}
\usage{
add_relief(
  classified_pts,
  crosswalk_id = NULL,
  pct_of_length_for_relief = 0.01
)
}
\arguments{
\item{classified_pts}{sf or dataframe of points with "hy_id", "cs_id", "cs_lengthm", and "point_type" columns. Output of hydrofabric3D::classify_points()}

\item{crosswalk_id}{character, ID column}

\item{pct_of_length_for_relief}{numeric, percent of cs_lengthm to use as the threshold depth for classifying whether a cross section has "relief". Default is 0.01 (1\% of the cross sections length).}
}
\value{
sf or dataframe with added "has_relief" columns or a dataframe of dataframe of unique hy_id/cs_id and "has_relief"
}
\description{
Add relief attributes to a dataframe of cross sections points
Given a set of cross section points (derived from hydrofabric3D::cross_section_pts() and hydrofabric3D::classify_points()) add a "has_relief" logical
value to data. The "has_relief" value is indicating whether a cross section "has relief".
Relief is determined by checking each set of cross section points have a left OR right bank that
has a depth difference from the bottom that isgreater than or equal to a percentage of the cross section length (e.g. Assuming a 'pct_of_length_for_relief' of 0.01 (1\%) of a 100m cross section would have a relief depth threshold of 1m)
}
