% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_without_network.R
\name{adjust_flagged_transects}
\alias{adjust_flagged_transects}
\title{Update a flagged set of transects by shortening them by the given extension_distance}
\usage{
adjust_flagged_transects(x, crosswalk_id = NULL, reindex_cs_ids = FALSE)
}
\arguments{
\item{x}{sf dataframe of transects}

\item{crosswalk_id}{character, unique ID column}

\item{reindex_cs_ids}{logical, whether to generate a new 1-n set of cs_ids or to return the original identifiers}
}
\value{
sf dataframe of transects with updated geometries
}
\description{
Update a flagged set of transects by shortening them by the given extension_distance
}
