% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_without_network.R
\name{adjust_flagged_transects2}
\alias{adjust_flagged_transects2}
\title{Update a flagged set of transects by shortening them by the given left_distance and right_distance (adjust_flagged_transects() v2 )
Requires 'left_distance' and 'right_distance' columns to specify how much to adjust flagged transects by}
\usage{
adjust_flagged_transects2(x, crosswalk_id = NULL, reindex_cs_ids = FALSE)
}
\arguments{
\item{x}{sf dataframe of transects}

\item{crosswalk_id}{character, unique ID column}

\item{reindex_cs_ids}{logical, whether to generate a new 1-n set of cs_ids or to return the original identifiers}
}
\value{
sf dataframe of transects with updated geometries
}
\description{
Update a flagged set of transects by shortening them by the given left_distance and right_distance (adjust_flagged_transects() v2 )
Requires 'left_distance' and 'right_distance' columns to specify how much to adjust flagged transects by
}
