% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_without_network.R
\name{adjust_transect_lengths}
\alias{adjust_transect_lengths}
\title{Extend/shrink an sf linestring dataframe by a specified lengths vector}
\usage{
adjust_transect_lengths(
  x,
  crosswalk_id = NULL,
  dir = "left",
  length_col = NULL
)
}
\arguments{
\item{x}{linestring sf dataframe, requires an}

\item{crosswalk_id}{character, unique ID column name}

\item{dir}{direction to extend/shrink transect from, either "left" or "right". Default is "left".}

\item{length_col}{character, name of the column in "x" that has the length of the linestring (meters)}
}
\value{
sf dataframe with extended linestring geometries
}
\description{
Extend/shrink an sf linestring dataframe by a specified lengths vector
}
