% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_banks_and_bottoms.R
\name{align_banks_and_bottoms}
\alias{align_banks_and_bottoms}
\title{Align banks and smooth bottoms of cross section points}
\usage{
align_banks_and_bottoms(cs_pts, crosswalk_id)
}
\arguments{
\item{cs_pts}{dataframe or sf dataframe of classified cross section points (output of classify_points())}

\item{crosswalk_id}{character, ID column}
}
\value{
sf dataframe of cross section points with aligned banks and smoothed bottoms
}
\description{
Ensures the bottom of each cross section is lower then or equal to that one upstream.
To do this, we traverse down the network making sure this condition is met, and,
in cases where it isn't, we will lower the in channel portion of the cross section to make it true.
}
