% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{bfs_traversal}
\alias{bfs_traversal}
\title{Traverse graph in a Breadth First Search (BFS) manner}
\usage{
bfs_traversal(network, start = NULL, return_as = "vector", verbose = FALSE)
}
\arguments{
\item{network}{data.frame with ID, toID, hydroseq, startflag, terminalpa, and divergence attributes.}

\item{start}{COMID of node to start traversal and go upstream/downstream from}

\item{return_as}{character string of how to return DFS traversal output. Either "list", "dataframe", "vector". Default is "vector", which return a list of 2 vectors, one indicating the BFS order of nodes and the other the respective level of each node.}

\item{verbose}{logical print status updates?}
}
\value{
data.frame containing the distance between pairs of network outlets.
}
\description{
Given a network with ID and toID columns, and a starting ID, traverse a river network using a Breadth First Search (BFS) algorithm to explore all nodes of directed acyclic graph (DAG)
}
