% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{braided_flowlines}
\alias{braided_flowlines}
\title{Braided Flowlines}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 508 rows and 39 columns.
}
\source{
Generated using hydrofabric3D software.
}
\usage{
braided_flowlines
}
\description{
A dataset containing flowlines representing braided river sections.
These flowlines are used in hydrologic models to simulate complex river networks.
}
\keyword{datasets}
