% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_utils.R
\name{calc_extension_distances}
\alias{calc_extension_distances}
\title{Calculate the minimum distance a line would need to extend to reach the boundary of the polygon/line that the input geometries are entirely within}
\usage{
calc_extension_distances(
  geos_geoms,
  ids,
  lines_to_cut,
  lines_to_cut_indices,
  direction = "head",
  max_extension_distance = 2500
)
}
\arguments{
\item{geos_geoms}{list of geos_geometrys}

\item{ids}{character vector}

\item{lines_to_cut}{geos_linestrings}

\item{lines_to_cut_indices}{numeric vector}

\item{direction}{character, either "head", "tail" or "both"}

\item{max_extension_distance}{numeric}
}
\value{
numeric vecotr, distance to extend each geos_geoms
}
\description{
Calculate the minimum distance a line would need to extend to reach the boundary of the polygon/line that the input geometries are entirely within
}
