% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankful_width.R
\name{calc_powerlaw_bankful_width}
\alias{calc_powerlaw_bankful_width}
\title{Calculates the powerlaw derived Bankful width given total drainage area (square kilometers)}
\usage{
calc_powerlaw_bankful_width(total_drainage_area_sqkm)
}
\arguments{
\item{total_drainage_area_sqkm}{}
}
\value{
numeric, bankful width estimate
}
\description{
Calculates the powerlaw derived Bankful width given total drainage area (square kilometers)
}
