% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validity_scores.R
\name{calc_validity_scores}
\alias{calc_validity_scores}
\title{Calculates a validity score column based on valid_banks and has_relief columns in a set of cross section points}
\usage{
calc_validity_scores(
  cs_to_validate,
  crosswalk_id = NULL,
  validity_col_name = "validity_score"
)
}
\arguments{
\item{cs_to_validate}{dataframe}

\item{crosswalk_id}{character, ID column}

\item{validity_col_name}{name of the output validity score column}
}
\value{
dataframe with added validity_score column
}
\description{
Calculates a validity score column based on valid_banks and has_relief columns in a set of cross section points
}
