% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{check_z_values}
\alias{check_z_values}
\title{Check for any Z values that are all equal or within a given threshold value}
\usage{
check_z_values(pts, threshold = 1)
}
\arguments{
\item{pts}{sf points dataframe}

\item{threshold}{numeric, default is 1 meter}
}
\value{
dataframe with hy_id, cs_id, Z, and is_same_Z value columns
}
\description{
Check for any Z values that are all equal or within a given threshold value
}
