% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{classify_banks_and_bottoms}
\alias{classify_banks_and_bottoms}
\title{Classify banks and bottoms}
\usage{
classify_banks_and_bottoms(num_of_pts, pt_ids, depths)
}
\arguments{
\item{num_of_pts}{integer}

\item{pt_ids}{numeric vector}

\item{depths}{numeric vector}
}
\value{
character vector
}
\description{
Classify banks and bottoms
}
