% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{classify_derivatives}
\alias{classify_derivatives}
\title{Classify a numeric vector based on 1st and 2nd derivative values}
\usage{
classify_derivatives(depths)
}
\arguments{
\item{depths}{numeric vector}
}
\value{
character vector
}
\description{
Given depths of a cross section, calculate the first and second derivative of the points
and then classify the points based on 1st / 2nd derivative values at each point.
With the intention that points with concavity / linear should be "channel" points
Points are labeled as:
\itemize{
\item flat
\item concave_up_increasing
\item concave_up_decreasing
\item concave_down_increasing
\item concave_down_decreasing
\item "linear"
}
}
