% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{classify_points}
\alias{classify_points}
\title{Classify Cross Section Points}
\usage{
classify_points(cs_pts)
}
\arguments{
\item{cs_pts}{CS points}
}
\value{
sf object
}
\description{
Classify Cross Section Points
}
