% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{classify_points4}
\alias{classify_points4}
\title{Classify Cross Section Points v1 (Deprecated version)}
\usage{
classify_points4(cs_pts)
}
\arguments{
\item{cs_pts}{CS points}
}
\value{
sf object
}
\description{
Classify Cross Section Points v1 (Deprecated version)
}
