% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{clean_point_types}
\alias{clean_point_types}
\title{Cleanup some common invalid point type classifications in cross section points
given a vector of point types representing a cross section of depths, set rules on what points can exist in proximity to other points
\itemize{
\item set any banks surrounded by channel and bottoms to "channel"
\item set any bank points surrounded by "channel" points to "channel"
\item set any banks surrounded by bottom points to "bottom"
}}
\usage{
clean_point_types(point_types)
}
\arguments{
\item{point_types}{character vector}
}
\value{
character vector
}
\description{
Cleanup some common invalid point type classifications in cross section points
given a vector of point types representing a cross section of depths, set rules on what points can exist in proximity to other points
\itemize{
\item set any banks surrounded by channel and bottoms to "channel"
\item set any bank points surrounded by "channel" points to "channel"
\item set any banks surrounded by bottom points to "bottom"
}
}
