% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_dissolve}
\alias{cluster_dissolve}
\title{Dissove polygons based on intersections with other polygons}
\usage{
cluster_dissolve(x)
}
\arguments{
\item{x}{sf dataframe}
}
\value{
dissolved sf dataframe
}
\description{
Dissove polygons based on intersections with other polygons
}
