% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validity_scores.R
\name{compare_cs_validity}
\alias{compare_cs_validity}
\title{Compare valid_banks and has_relief between 2 sets of cross section points}
\usage{
compare_cs_validity(cs_pts1, cs_pts2, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts1}{dataframe or sf dataframe of CS pts}

\item{cs_pts2}{dataframe or sf dataframe of CS pts}

\item{crosswalk_id}{character unique ID}
}
\value{
dataframe, tibble
}
\description{
Compare valid_banks and has_relief between 2 sets of cross section points
}
