% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_ordering.R
\name{cs_arrange}
\alias{cs_arrange}
\title{Rearrange transects / cross sections in order from upstream to downstream}
\usage{
cs_arrange(x, crosswalk_id = NULL, order_by = c("cs_id", "cs_measure"))
}
\arguments{
\item{x}{dataframe, sf dataframe or tibble}

\item{crosswalk_id}{character, unique ID column}

\item{order_by}{character, either "cs_id" or "cs_measure"}
}
\value{
dataframe, sf dataframe or tibble with an added 'cs_id' column
}
\description{
Rearrange transects / cross sections in order from upstream to downstream
}
