% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahg_estimates.R
\name{cs_inchannel_estimate}
\alias{cs_inchannel_estimate}
\title{Given provide inchannel widths and depths to a set of cross section points and derive estimated shapes}
\usage{
cs_inchannel_estimate(
  cross_section_pts,
  r = 3,
  inchannel_width,
  inchannel_depth,
  drop_negative_depths = FALSE
)
}
\arguments{
\item{cross_section_pts}{dataframe or sf dataframe}

\item{r}{numeric, R coefficient}

\item{inchannel_width}{numeric}

\item{inchannel_depth}{numeric}

\item{drop_negative_depths}{logical, whether to remove any depths that are negative, default is FALSE (probably Deprecated at this point)}
}
\value{
dataframe or sf dataframe with AHG estimated points injected into the input cross section points
}
\description{
Still in early development phases
}
