% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{cut_cross_sections}
\alias{cut_cross_sections}
\title{Generate Cross Sections Across Hydrographic Network}
\usage{
cut_cross_sections(
  net,
  id = NULL,
  bf_widths = 100,
  num = 10,
  smooth = TRUE,
  densify = FALSE
)
}
\arguments{
\item{net}{Hydrographic LINESTRING Network}

\item{id}{Uniuqe Identifier in net}

\item{bf_widths}{Bankfull Widths (length of cross sections for each net element)}

\item{num}{Number of transects per Net element}
}
\value{
sf object
}
\description{
Generate Cross Sections Across Hydrographic Network
}
