% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{cut_transect}
\alias{cut_transect}
\title{Generate a Perpendicular Linestring of a Given Width}
\usage{
cut_transect(edge, width)
}
\arguments{
\item{edge}{LINESRTING}

\item{width}{Length of Perpendicular LINESTRING}
}
\value{
GEOS object
}
\description{
Generate a Perpendicular Linestring of a Given Width
}
