% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{cycle_edge_list_to_cycle_edge_df}
\alias{cycle_edge_list_to_cycle_edge_df}
\title{Convert output of find_cycles() to a dataframe
Converts a node cycle list to a dataframe of nodes and cycle IDs, internal helper function.}
\usage{
cycle_edge_list_to_cycle_edge_df(edge_cycles_list, wide = FALSE)
}
\arguments{
\item{edge_cycles_list}{list of edges in each cycle_id (output of find_cycles())}

\item{wide}{logical, return data as a wide dataframe (multiple cycle_ids in a list in each row), or long dataframe (1 cycle_id for each row representing a node)}
}
\value{
tibble, dataframe
}
\description{
Convert output of find_cycles() to a dataframe
Converts a node cycle list to a dataframe of nodes and cycle IDs, internal helper function.
}
