% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{dfs_traversal}
\alias{dfs_traversal}
\title{Traverse graph in a Depth First Search (DFS) manner}
\usage{
dfs_traversal(network, start = NULL, return_as = "list", verbose = FALSE)
}
\arguments{
\item{network}{data.frame with ID, toID, hydroseq, startflag, terminalpa, and divergence attributes.}

\item{start}{COMID of node to start traversal and go upstream/downstream from}

\item{return_as}{character string of how to return DFS traversal output. Either "list", "dataframe", "vector". Default is "list".}

\item{verbose}{logical print status updates?}
}
\value{
data.frame containing the distance between pairs of network outlets.
}
\description{
Given a network with ID and toID columns, and a starting ID, traverse a river network using a Depth First Search (DFS) algorithm to explore all nodes of directed acyclic graph (DAG)
}
