% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_overlaps}
\alias{diff_overlaps}
\title{Use sf::st_difference to resolve overlaps in polygons based on intersections with other polygons}
\usage{
diff_overlaps(x)
}
\arguments{
\item{x}{sf dataframe}
}
\value{
sf dataframe with overlaps removed
}
\description{
Use sf::st_difference to resolve overlaps in polygons based on intersections with other polygons
}
