% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_incomplete_cs_pts}
\alias{drop_incomplete_cs_pts}
\title{Remove entire cross sections that have any NA Z (depth) values}
\usage{
drop_incomplete_cs_pts(cross_section_pts, crosswalk_id = NULL)
}
\arguments{
\item{cross_section_pts}{cs points dataframe, tibble, or sf dataframe}

\item{crosswalk_id}{unique ID for flowline}
}
\value{
cross_section_pts dataframe / tibble / sf dataframe with removed cross sections
}
\description{
Remove entire cross sections that have any NA Z (depth) values
}
