% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{extend_by_length}
\alias{extend_by_length}
\title{Extend an sf linestring dataframe by a specified lengths vector}
\usage{
extend_by_length(x, length_vector, length_col = NULL)
}
\arguments{
\item{x}{linestring sf dataframe}

\item{length_vector}{numeric, vector of length 'x' representing the number of meters to extend 'x' from both directions (i.e. 10 means the linestring will be extended 10m from both ends of the line)}

\item{length_col}{character, name of the column in "x" that has the length of the linestring (meters)}
}
\value{
sf dataframe with extended linestring geometries
}
\description{
Extend an sf linestring dataframe by a specified lengths vector
}
