% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders.R
\name{extend_by_percent}
\alias{extend_by_percent}
\title{Extend an sf linestring dataframe by a percent of the lines length}
\usage{
extend_by_percent(x, pct = 0.5, length_col = NULL)
}
\arguments{
\item{x}{linestring sf dataframe}

\item{pct}{numeric, percent of line to extend linestring by in both directions}

\item{length_col}{character, name of the column in "x" that has the length of the linestring (meters)}
}
\value{
sf dataframe with extended linestring geometries
}
\description{
Extend an sf linestring dataframe by a percent of the lines length
}
