% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_with_network.R
\name{extend_invalid_transects3}
\alias{extend_invalid_transects3}
\title{Extend a set of transects by a percentage based on banks and relief
Given a set of transect lines with valid_banks and has_relief columns (derived from DEM extracted cross section points), extend any transects
by a percentage of the transects length if the transect does NOT have valid banks (valid_banks == FALSE) OR it does NOT have relief (has_relief == FALSE).}
\usage{
extend_invalid_transects3(transects_to_check, net, scale = 0.5, verbose = TRUE)
}
\arguments{
\item{transects_to_check}{sf linestrings, set of all transects in the network. Requires the following columns: "hy_id", "cs_id", "cs_lengthm" (length of geometry in meters), "valid_banks", and "has_relief"}

\item{net}{sf linestrings, flowline network that transects were generated from, requires "id" column (where "id" equals the "hy_id" columns in 'transects_to_check' and 'transects' )}

\item{scale}{numeric, percentage of current transect line length to extend transects in transects_to_extend by. Default is 0.5 (50\% of the transect length)}

\item{verbose}{logical, whether to show a progress bar and progress messages or not. Default is TRUE.}
}
\value{
sf linestring dataframe containing the the original transects with extensions performed on transects without valid_banks OR has_relief (a "is_extended" flag denotes if the geometry was extended by "scale" \% or not)
}
\description{
Extend a set of transects by a percentage based on banks and relief
Given a set of transect lines with valid_banks and has_relief columns (derived from DEM extracted cross section points), extend any transects
by a percentage of the transects length if the transect does NOT have valid banks (valid_banks == FALSE) OR it does NOT have relief (has_relief == FALSE).
}
