% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_based_transect_extenders.R
\name{extend_transects_any_side}
\alias{extend_transects_any_side}
\title{Given a set of transect lines, a flowline network, extend the transect lines out given distances from the left and right
Flowlines are required to ensure valid transect intersection relationship is maintained}
\usage{
extend_transects_any_side(
  transects,
  flowlines,
  crosswalk_id,
  cs_id = "cs_id",
  grouping_id = "mainstem"
)
}
\arguments{
\item{transects}{sf dataframe of linestrings, requires crosswalk_id, cs_id, grouping_id columns and numeric 'left_distance' and 'right_distance' columns}

\item{flowlines}{sf dataframe of linestrings}

\item{crosswalk_id}{character, column name that connects features in transects to flowlines}

\item{cs_id}{character, column name that uniquely identifies transects within a flowline}

\item{grouping_id}{character, column name in both transects and flowlines that denotes which flowlines are grouped with which transects.}
}
\value{
transects sf dataframe with extended transect geometries, left and right distance columns, and flags indicating if the transect was extended in the left and/or right directions
}
\description{
Given a set of transect lines, a flowline network, extend the transect lines out given distances from the left and right
Flowlines are required to ensure valid transect intersection relationship is maintained
}
