% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_with_network.R
\name{extend_transects_by_cs_attributes}
\alias{extend_transects_by_cs_attributes}
\title{Extend transects for any transects with invalid cross section attributes}
\usage{
extend_transects_by_cs_attributes(
  transects = NULL,
  flowlines = NULL,
  crosswalk_id = NULL,
  scale = 0.5,
  keep_lengths = FALSE,
  reindex_cs_ids = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{transects}{sf dataframe of transect LINESTRING geometries}

\item{flowlines}{sf dataframe of flowline LINESTRING geometries}

\item{crosswalk_id}{character}

\item{scale}{numeric percent of original transect length to extend (in both directions). Default is 0.5 or 50\% of transects length (i.e. 25\% increase in length in both directions).}

\item{keep_lengths}{logical whether to keep a record of the original transect lengths or not, default is FALSE, original lengths are not kept}

\item{reindex_cs_ids}{logical, whether to reindex the cs_ids to ensure each crosswalk_id has cs_ids of 1-number of transects. Default is FALSE, which guarantees
crosswalk_id/cs_ids remain untouched as they were given in the input data. Setting this to TRUE will make sure if any cross sections were removed from a crosswalk_id, then the cs_ids are renumbered so there are no gaps between cs_ids within a crosswalk_id}

\item{verbose}{logical}
}
\value{
dataframe or sf dataframe of extended transects
}
\description{
Extend transects for any transects with invalid cross section attributes
}
