% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders.R
\name{extend_transects_by_length}
\alias{extend_transects_by_length}
\title{Extend a set of transects by a percentage}
\usage{
extend_transects_by_length(
  transects_to_extend,
  length_vector,
  net,
  verbose = TRUE
)
}
\arguments{
\item{transects_to_extend}{sf linestrings, set of all transects in the network. Requires the following columns: "hy_id", "cs_id", "cs_lengthm" (length of geometry in meters),}

\item{length_vector}{numeric, vector of length 'x' representing the number of meters to extend 'x' from both directions (i.e. 10 means the linestring will be extended 10m from both ends of the line)}

\item{net}{sf linestrings, flowline network that transects were generated from, requires "id" column (where "id" equals the "hy_id" columns in 'transects_to_extend' and 'transects' )}

\item{verbose}{logical, whether to print messages or not. Default is TRUE}
}
\value{
sf linestring dataframe containing the updates transects_to_extend (with a flag denoting if the geometry was extended by "scale" \% or not)
}
\description{
Extend a set of transects by a percentage
}
