% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_with_network.R
\name{extend_transects_sides}
\alias{extend_transects_sides}
\title{Given a set of transect lines, a flowline network, extend the transect lines out given distances from the left and right
Flowlines are required to ensure valid transect intersection relationship is maintained}
\usage{
extend_transects_sides(
  transects,
  flowlines,
  crosswalk_id,
  cs_id = "cs_id",
  grouping_id = "mainstem",
  direction = "any"
)
}
\arguments{
\item{transects}{sf dataframe of linestrings, requires crosswalk_id, cs_id, grouping_id columns and numeric 'extension_distance' column indicating the distance to extend}

\item{flowlines}{sf dataframe of linestrings}

\item{crosswalk_id}{character, column name that connects features in transects to flowlines}

\item{cs_id}{character, column name that uniquely identifies transects within a flowline}

\item{grouping_id}{character, column name in both transects and flowlines that denotes which flowlines are grouped with which transects.}

\item{direction}{character, whether to extend transects individually from left and right sides, or to strictly extend a transect if BOTH the left and right extension are valid. Valid inputs are be either "any" or "both".}
}
\value{
transects sf dataframe with extended transect geometries, left and right distance columns, and flags indicating if the transect was extended in the left and/or right directions
}
\description{
Given a set of transect lines, a flowline network, extend the transect lines out given distances from the left and right
Flowlines are required to ensure valid transect intersection relationship is maintained
}
