% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{extract_dem_values}
\alias{extract_dem_values}
\title{Given a set of linestrings, extract DEM values at points along the linestring}
\usage{
extract_dem_values(cs, dem)
}
\arguments{
\item{cs}{cross section sf object}

\item{dem}{SpatRaster DEM or character pointing to remote DEM resource}
}
\value{
sf dataframe with Z values extracted from DEM
}
\description{
Given a set of linestrings, extract DEM values at points along the linestring
}
