% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{extract_pt_val}
\alias{extract_pt_val}
\title{Extract raster values at points}
\usage{
extract_pt_val(rast, pts)
}
\arguments{
\item{rast}{terra SpatRaster}

\item{pts}{sf or terra points}
}
\value{
numeric vector of values from SpatRaster
}
\description{
Extract raster values at points
}
