% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahg_estimates.R
\name{fill_missing_ahg_coords}
\alias{fill_missing_ahg_coords}
\title{Generate X/Y coordinates between a set of known points within a cross section
Used after inserting AHG estimated parabolas in between DEM cross sections points}
\usage{
fill_missing_ahg_coords(cross_section_pts)
}
\arguments{
\item{cross_section_pts}{cross section points dataframe with missing X/Y coordinates between sets of known X/Y coordinates}
}
\value{
dataframe, input dataframe with X/Y coordinates filled in for missing hy_id/cs_id X/Y values
}
\description{
Generate X/Y coordinates between a set of known points within a cross section
Used after inserting AHG estimated parabolas in between DEM cross sections points
}
