% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{find_anchor_pts}
\alias{find_anchor_pts}
\title{Find the anchor points on the left and right of the bottoms (where bottom points end)
given depths and relative distances and points classified into banks vs bottoms
determine the indices for
\itemize{
\item start of the left side bank/channel
\item start of the right/side bank channel
\item the index of the point at the middle of the bottom
}}
\usage{
find_anchor_pts(depths, num_of_pts, cs_length, relative_distance, point_types)
}
\arguments{
\item{depths}{}

\item{num_of_pts}{}

\item{cs_length}{}

\item{relative_distance}{}

\item{point_types}{}
}
\value{
numeric vector of length 3 with the index of the left anchor, middle point, and right anchor (i.e. c(2, 5, 9) -> c(left_anchor, middle_bottom, right_anchor))
}
\description{
Find the anchor points on the left and right of the bottoms (where bottom points end)
given depths and relative distances and points classified into banks vs bottoms
determine the indices for
\itemize{
\item start of the left side bank/channel
\item start of the right/side bank channel
\item the index of the point at the middle of the bottom
}
}
