% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_area.R
\name{find_cs_area}
\alias{find_cs_area}
\title{Given a specific depth, an array of depths, and relative_distances return the cross sectional area}
\usage{
find_cs_area(depth, depth_array, relative_distance)
}
\arguments{
\item{depth}{numeric, depth at a specific point along CS}

\item{depth_array}{numeric set of all depths for CS}

\item{relative_distance}{numeric vector of distances from 0 to length of CS}
}
\value{
numeric cross sectional area value, 0 if no area can be calculated
}
\description{
Given a specific depth, an array of depths, and relative_distances return the cross sectional area
}
