% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_transects.R
\name{fix_braid_transects2}
\alias{fix_braid_transects2}
\title{Fix transects found on braided river sections (v2)}
\usage{
fix_braid_transects2(
  net,
  transect_lines,
  crosswalk_id = NULL,
  method = "comid",
  precision = 1,
  rm_intersects = TRUE
)
}
\arguments{
\item{net}{sf dataframe of hydrologic network, linestrings}

\item{transect_lines}{sf linestring dataframe, containing cross sections of flowlines in 'net' the output of "cut_cross_sections2()" function}

\item{crosswalk_id}{character, unique ID column}

\item{method}{The method to determine the geometries to cut. Options are "comid", "component", or "neighbor". Default is "comid"}

\item{precision}{int, distance in meters to approximate final cross section linestring length. Value you must be greater than 0. Default is 1}

\item{rm_intersects}{logical, whether to remove transect linestrings that intersect with other parts of the network ('net'). Default is TRUE which will remove intersecting linestrings.}
}
\value{
sf object of transect linestrings
}
\description{
Fix transects found on braided river sections (v2)
}
