% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_braided_transects.R
\name{fix_braided_transects}
\alias{fix_braided_transects}
\title{Fix transects found on braided river sections (latest)}
\usage{
fix_braided_transects(
  network,
  transect_lines,
  crosswalk_id = NULL,
  braid_threshold = NULL,
  method = "crosswalk_id",
  precision = 1,
  rm_intersects = TRUE
)
}
\arguments{
\item{network}{sf dataframe of hydrologic network, linestrings}

\item{transect_lines}{sf linestring dataframe, containing cross sections of flowlines in 'network' the output of "cut_cross_sections()" function}

\item{crosswalk_id}{character, unique ID column}

\item{braid_threshold}{numeric value, value of the total length of all flowlines in a braid. Only braids with total flowline
lengths less than or equal to the threshold will be considered by function (i.e. determines that maximum braid size that fix_braid_transects() should operate on).
Default is NULL, which will attempt to fix all the braid transects in the data}

\item{method}{The method to determine the geometries to cut. Options are "crosswalk_id", "component", or "neighbor". Default is "crosswalk_id"}

\item{precision}{int, distance in meters to approximate final cross section linestring length. Value you must be greater than 0. Default is 1}

\item{rm_intersects}{logical, whether to remove transect linestrings that intersect with other parts of the network ('network'). Default is TRUE which will remove intersecting linestrings.}
}
\value{
sf object of transect linestrings
}
\description{
Fix transects found on braided river sections (latest)
}
