% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahg_estimates.R
\name{fix_oversized_topwidths}
\alias{fix_oversized_topwidths}
\title{Check that all cross sections points have a prescribed top width less than the total cross section length}
\usage{
fix_oversized_topwidths(cross_section_pts = NULL, crosswalk_id = NULL)
}
\arguments{
\item{cross_section_pts}{dataframe or sf dataframe with "hy_id", "cs_id", "pt_id", "Z", "relative_distance", "cs_lengthm", "class", "point_type", "TW", "DEPTH", "DINGMAN_R"}

\item{crosswalk_id}{character, ID column}
}
\value{
cross_section_pts dataframe with updated "top_width" and "depth" column values
}
\description{
If a set of cross section points has a top width length that is longer than the cross sections length, then a new top width and Y max (depth) value
are given so that the estimated shape is able to properly fit into the cross sections.
The cross sections length (meters) minus 1 meter is used as the new top width and
the new Y max (depth) value is derived from the original ratio between the prescribed top width and Y max
}
