% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_improvements.R
\name{flag_transects_for_change}
\alias{flag_transects_for_change}
\title{Add a flagged and extension distance columns to set of transects with CS attributes based on new cross section points data}
\usage{
flag_transects_for_change(
  x,
  crosswalk_id = NULL,
  points_per_cs = NULL,
  min_pts_per_cs = 10,
  dem =
    "/vsicurl/https://prd-tnm.s3.amazonaws.com/StagedProducts/Elevation/13/TIFF/USGS_Seamless_DEM_13.vrt",
  pct_of_length_for_relief = 0.01
)
}
\arguments{
\item{x}{sf dataframe of transects}

\item{crosswalk_id}{character, unique ID column}

\item{points_per_cs}{numeric}

\item{min_pts_per_cs}{numeric}

\item{dem}{character}

\item{pct_of_length_for_relief}{numeric}
}
\value{
sf dataframe of transects with updated geometries
}
\description{
Add a flagged and extension distance columns to set of transects with CS attributes based on new cross section points data
}
