% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowlines}
\alias{flowlines}
\title{Flowlines}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\source{
NOAA Office of Water Prediction.
}
\usage{
flowlines
}
\description{
A dataset of primary flowlines for hydrologic and hydraulic modeling.
}
\keyword{datasets}
