% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowlines_missing_depth}
\alias{flowlines_missing_depth}
\title{Flowlines Missing Depth}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 5 columns.
}
\source{
Derived from \code{flowlines}.
}
\usage{
flowlines_missing_depth
}
\description{
A dataset of flowlines missing depth information, which may require further analysis or imputation.
}
\keyword{datasets}
