% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_transects.R
\name{geos_extend_line}
\alias{geos_extend_line}
\title{Extend a geos_geometry linestring from, one or both ends, by a given distance (meters)}
\usage{
geos_extend_line(line, distance, dir = "both", with_crs = TRUE)
}
\arguments{
\item{line}{sf linestring or geos_geometry linestring to extend}

\item{distance}{numeric value in meters or a vector of length 2 if 'end = "both"' where}

\item{dir}{character, determines whether to extend the linestring from the 'tail', 'head' or 'both' ends}

\item{with_crs}{logical, whether a CRS should be prescribed to extended output geos_geometry linestring}
}
\value{
geos_geometry linestring extended by 'distance' from either the 'head', 'tail' or 'both' ends of the original linestring
}
\description{
Extend a geos_geometry linestring from, one or both ends, by a given distance (meters)
}
