% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_bank_attributes.R
\name{get_bank_attributes}
\alias{get_bank_attributes}
\title{Get attributes about the banks of each cross section in a dataframe of cross section points
Given a set of cross section points with point_type column, return a dataframe of the unique hy_id/cs_ids with the following calculated columns:
"bottom", "left_bank", "right_bank" columns which are the Z values of the "lowest" bottom point, and the "highest" left and right bank Z values, respectively.
And a "valid_banks" column indicating whether the hy_id/cs_id set of cross section point has at least a signle bottom point with
at least 1 left bank point AND 1 right bank point that are above the lowest "bottom" point.}
\usage{
get_bank_attributes(classified_pts, crosswalk_id = NULL)
}
\arguments{
\item{classified_pts}{sf or dataframe of points with "hy_id", "cs_id", and "point_type" columns. Output of hydrofabric3D::classify_pts()}

\item{crosswalk_id}{character, ID column}
}
\value{
dataframe with each row being a unique hy_id/cs_id with "bottom", "left_bank", "right_bank", and "valid_banks" values for each hy_id/cs_id.
}
\description{
Get attributes about the banks of each cross section in a dataframe of cross section points
Given a set of cross section points with point_type column, return a dataframe of the unique hy_id/cs_ids with the following calculated columns:
"bottom", "left_bank", "right_bank" columns which are the Z values of the "lowest" bottom point, and the "highest" left and right bank Z values, respectively.
And a "valid_banks" column indicating whether the hy_id/cs_id set of cross section point has at least a signle bottom point with
at least 1 left bank point AND 1 right bank point that are above the lowest "bottom" point.
}
