% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{get_braid_list}
\alias{get_braid_list}
\title{Create a list of braid IDs containing COMIDs in each braid}
\usage{
get_braid_list(network, terminal_id = NULL, verbose = FALSE)
}
\arguments{
\item{network}{The network object representing the river network.}

\item{terminal_id}{character, column name containing a unique identifier, delineating seperate networks in the 'network' dataset. Default is NULL which will use 'find_connected_components()' and determine the connected components in the graph to try and create a 'component_id' column in 'network'}

\item{verbose}{Logical indicating whether to display verbose messages during the braid detection process.}
}
\value{
list of braid IDs and COMIDs within each braid
}
\description{
Find and uniquely identify braids in a network of flowlines, given a dataframe containing comid, fromnode, tonode and divergence as columns. 'find_braids()" identifies braids as cycles in the graph representation of the river network.
}
\examples{
 \dontrun{
net <- nhdplusTools::navigate_network(
 start       = 101,
 mode        = "UT",
 distance_km = 100
 ) 

net <- dplyr::select(net, comid, divergence, totdasqkm, fromnode, tonode, terminalpa)

# get a dataframe of COMIDs and braid IDs
braids <- get_braid_list(network = net)
}
}
