% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahg_estimates.R
\name{get_coords_around_parabola}
\alias{get_coords_around_parabola}
\title{Get the coordinates surrounding a set of missing AHG X/Y coordinates.}
\usage{
get_coords_around_parabola(cross_section_pts)
}
\arguments{
\item{cross_section_pts}{dataframe with cross section points, (required cols, "hy_id", "cs_id", "X", "Y", "is_dem_point")}
}
\value{
dataframe with each hy_id/cs_id cross section containing a value for X_start, X_end, Y_start, Y_end, representing the points surrounding the AHG inserted points
}
\description{
Get the coordinates surrounding a set of missing AHG X/Y coordinates.
}
