% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cs_bottom_length}
\alias{get_cs_bottom_length}
\title{Calculate the length between the leftmost and rightmost bottom point in each cross section}
\usage{
get_cs_bottom_length(cross_section_pts)
}
\arguments{
\item{cross_section_pts}{}
}
\value{
summarized dataframe of input cross_section_pts dataframe with a bottom_length value for each hy_id/cs_id
}
\description{
Calculate the length between the leftmost and rightmost bottom point in each cross section
}
