% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_utils.R
\name{get_line_node_pts}
\alias{get_line_node_pts}
\title{Get the starting and ending points of geos_linestring
Internal helper function}
\usage{
get_line_node_pts(line)
}
\arguments{
\item{extension}{geos_geometry linestring}

\item{use_extension}{logical, do we use the extension}
}
\value{
geos_geometry points, the start and end point of the final extension line
}
\description{
Get the starting and ending points of geos_linestring
Internal helper function
}
