% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_utils.R
\name{get_lines_extended_to_geoms_subset}
\alias{get_lines_extended_to_geoms_subset}
\title{Calculate the minimum distance a line would need to extend to reach the boundary of the polygon/line that the input geometries are entirely within
Version of get_lines_extended_to_geoms() but only iterates through the IDs/geometries that are predetermined to be WITHIN A POLYGON}
\usage{
get_lines_extended_to_geoms_subset(
  geos_geoms,
  ids,
  lines_to_cut,
  lines_to_cut_indices,
  direction = "head",
  max_extension_distance = 2500,
  verbose = FALSE
)
}
\arguments{
\item{geos_geoms}{list of geos_geometrys}

\item{ids}{character vector}

\item{lines_to_cut}{geos_linestrings}

\item{lines_to_cut_indices}{numeric vector}

\item{direction}{character, either "head", "tail" or "both"}

\item{max_extension_distance}{numeric}

\item{verbose}{logical, whether to print messages or not. Default is FALSE}
}
\value{
geos_geometry vector of extended linestrings for the geometries within the lines to cut
}
\description{
Calculate the minimum distance a line would need to extend to reach the boundary of the polygon/line that the input geometries are entirely within
Version of get_lines_extended_to_geoms() but only iterates through the IDs/geometries that are predetermined to be WITHIN A POLYGON
}
