% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_point_type_counts}
\alias{get_point_type_counts}
\title{Get the count of each point type in a set of cross section points
classify_points() will add the required "point_type" column to a set of cross section points}
\usage{
get_point_type_counts(classified_pts, add = TRUE)
}
\arguments{
\item{classified_pts}{dataframe or sf dataframe, cross section points with a "hy_id", and "cs_id" columns as well asa 'point_type' column containing the values: "bottom", "left_bank", "right_bank", and "channel"}

\item{add}{logical, whether to add the point type columns to the original data or not. Default is TRUE. If FALSE, a dataframe with "hy_id", "cs_id", "left_bank_count", "right_bank_count", "channel_count", and ""bottom_count" is returned}
}
\value{
dataframe or sf dataframe
}
\description{
Get the count of each point type in a set of cross section points
classify_points() will add the required "point_type" column to a set of cross section points
}
