% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{get_points_per_cs}
\alias{get_points_per_cs}
\title{Calculate the points per cross section based off length}
\usage{
get_points_per_cs(
  cs_length,
  points_per_cs = NULL,
  min_pts_per_cs = 10,
  dem = default_dem
)
}
\arguments{
\item{cs_length}{numeric vector, lengths of each cross section (meters)}

\item{points_per_cs}{numeric, number of points per cross section. Default is NULL}

\item{min_pts_per_cs}{An optional minimum points value per cross section. If not provided,}

\item{dem}{A SpatRaster object representing the Digital Elevation Model (DEM) or a character string referencing a remote resource.
the function calculates it based on the length of cross-sections and the resolution of the DEM.}
}
\value{
numeric vector, indicating the number of points per cross section for each cs_length
}
\description{
Calculate the points per cross section based off length
}
