% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_relief}
\alias{get_relief}
\title{Get relief attributes from a dataframe of cross sections points
Generate a dataframe from a set of classified cross section points indicating whether a cross section "has relief".
Relief is determined by checking each set of cross section points have a left OR right bank that has a depth difference from the bottom that is
greater than or equal to a percentage of the cross section length (e.g. Assuming a 'pct_of_length_for_relief' of 0.01 (1\%) of a 100m cross section would have a relief depth threshold of 1m)}
\usage{
get_relief(
  classified_pts,
  crosswalk_id = NULL,
  pct_of_length_for_relief = 0.01,
  detailed = FALSE
)
}
\arguments{
\item{classified_pts}{sf or dataframe of points with "hy_id", "cs_id", "cs_lengthm", and "point_type" columns. Output of hydrofabric3D::classify_pts()}

\item{crosswalk_id}{character, ID column}

\item{pct_of_length_for_relief}{numeric, percent of cs_lengthm to use as the threshold depth for classifying whether a cross section has "relief". Default is 0.01 (1\% of the cross sections length).}

\item{detailed}{logical, whether to return only a the "has_relief" column or
include all derived relief based columns such as "max_relief" and the "pct_of_length_for_relief" used. Default is FALSE and returns a dataframe with only "hy_id", "cs_id", and "has_relief".}
}
\value{
dataframe with each row being a unique hy_id/cs_id with a "has_relief" value for each hy_id/cs_id. If detailed = TRUE, then the output dataframe will include the following additional columns: "cs_lengthm", "max_relief", "pct_of_length_for_relief".
}
\description{
Get relief attributes from a dataframe of cross sections points
Generate a dataframe from a set of classified cross section points indicating whether a cross section "has relief".
Relief is determined by checking each set of cross section points have a left OR right bank that has a depth difference from the bottom that is
greater than or equal to a percentage of the cross section length (e.g. Assuming a 'pct_of_length_for_relief' of 0.01 (1\%) of a 100m cross section would have a relief depth threshold of 1m)
}
