% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{get_start_node}
\alias{get_start_node}
\title{Get a valid starting node from a graph}
\usage{
get_start_node(graph, start = NULL)
}
\arguments{
\item{graph}{dataframe, sf dataframe, with fromnode and tonode columns}

\item{start}{character, node in 'fromnode' column of graph}
}
\value{
character, node
}
\description{
Get a valid starting node from a graph
}
