% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_to_polygon_extender.R
\name{get_transect_extension_distances_to_polygons}
\alias{get_transect_extension_distances_to_polygons}
\title{Get the left and right extension distances for a set of transects out to a set of polygons}
\usage{
get_transect_extension_distances_to_polygons(
  transects,
  polygons,
  crosswalk_id,
  max_extension_distance,
  tolerance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{transects}{sf linestring dataframe}

\item{polygons}{sf polygon dataframe}

\item{crosswalk_id}{character}

\item{max_extension_distance}{numeric}

\item{tolerance}{A minimum distance to use for simplification on polygons. Use a higher value for more simplification on the polygons. Default is NULL which will apply no simplification to polygons.}

\item{verbose}{logical, whether to output messages or not. Default is TRUE, and messages will output}
}
\value{
data.frame or tibble
}
\description{
Get the left and right extension distances for a set of transects out to a set of polygons
}
