% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{get_transects}
\alias{get_transects}
\title{Generate Multiple cross section along a linestring}
\usage{
get_transects(line, bf_width, n)
}
\arguments{
\item{line}{sf linestring or geos_geometry, original line element}

\item{bf_width}{Bankfull Width (length of cross section)}

\item{n}{number of cross sections}
}
\value{
sf dataframe with 'n' evenly spaced transect lines with cs_measures for each cross section geometry
}
\description{
Generate Multiple cross section along a linestring
}
