% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_tmp_ids}
\alias{get_unique_tmp_ids}
\title{Get a list of unique tmp_ids in a dataframe}
\usage{
get_unique_tmp_ids(df, x = "hy_id", y = "cs_id")
}
\arguments{
\item{df}{dataframe with x and y as columns, with an optional "tmp_id" column, otherwise a tmp_id will be created from x_y}

\item{x}{The name of the column in df to make up the first part of the added tmp_id column (tmp_id = x_y). Default is hy_id.}

\item{y}{The name of the column in df to make up the second part of the added tmp_id column (tmp_id = x_y). Default is cs_id.}
}
\value{
character vector of unique "tmp_id" values in the given dataframe
}
\description{
Dataframe can have "tmp_id" column already or the columns can be specified with 'x' and 'y' arguments
}
