% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invalid_flowlines}
\alias{invalid_flowlines}
\title{Invalid Flowlines}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.
}
\source{
Processed using \code{rm_self_intersections}.
}
\usage{
invalid_flowlines
}
\description{
A dataset of flowlines identified as invalid due to self-intersections or other topological errors.
}
\keyword{datasets}
