% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{is_bank_channel_bank_pattern}
\alias{is_bank_channel_bank_pattern}
\title{Check if a set of cross_section point types are in the pattern of left_bank, channel, then right_bank points}
\usage{
is_bank_channel_bank_pattern(point_types)
}
\arguments{
\item{point_types}{character vector}
}
\value{
logical, TRUE if points are left_bank --> channel --> right_bank order
}
\description{
Check if a set of cross_section point types are in the pattern of left_bank, channel, then right_bank points
}
