% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_sf_linestring}
\alias{is_sf_linestring}
\title{Check if data is an SF linestring / multilinestring}
\usage{
is_sf_linestring(data)
}
\arguments{
\item{data}{dataframe, tibble, sf dataframe, geometry}
}
\value{
logical, TRUE if data is an sf linestring / multilinestring
}
\description{
Check if data is an SF linestring / multilinestring
}
