% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_utils.R
\name{is_valid_transect_line}
\alias{is_valid_transect_line}
\title{Check if an updated transect line is valid relative to the other transects and flowlines in the network
The 'transect_to_check' should be 'used' (i.e. function returns TRUE) if
the 'transect_to_check' does NOT interesect any other transects ('transect_lines') AND it only intersects a single flowline ONCE.
If the 'transect_to_check' intersects ANY other transects OR intersects a flowline more
than once (OR more than one flowline in the network) then the function returns FALSE.}
\usage{
is_valid_transect_line(transect_to_check, transect_lines, flowlines)
}
\arguments{
\item{transect_to_check}{geos_geometry, linestring}

\item{transect_lines}{geos_geometry, linestring}

\item{flowlines}{geos_geometry, linestring}
}
\value{
TRUE if the extension should be used, FALSE if it shouldn't be used
}
\description{
Check if an updated transect line is valid relative to the other transects and flowlines in the network
The 'transect_to_check' should be 'used' (i.e. function returns TRUE) if
the 'transect_to_check' does NOT interesect any other transects ('transect_lines') AND it only intersects a single flowline ONCE.
If the 'transect_to_check' intersects ANY other transects OR intersects a flowline more
than once (OR more than one flowline in the network) then the function returns FALSE.
}
