% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{junction_flowlines}
\alias{junction_flowlines}
\title{Junction Flowlines}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 5 columns.
}
\source{
Derived from \code{flowlines}.
}
\usage{
junction_flowlines
}
\description{
A dataset of flowlines representing junctions in the river network, used for hydrodynamic connectivity analysis.
}
\keyword{datasets}
