% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/package.R
\docType{data}
\name{linestring}
\alias{linestring}
\title{Flowlines linestring}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 325 rows and 5 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 325 rows and 5 columns.
}
\source{
Generated using hydrofabric3D software.
}
\usage{
linestring

linestring
}
\description{
A dataset containing flowlines linestrings .
}
\keyword{datasets}
