% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_utils.R
\name{make_line_from_start_and_end_pts}
\alias{make_line_from_start_and_end_pts}
\title{Given 2 geos_geometry point geometries, create a line between the 2 points}
\usage{
make_line_from_start_and_end_pts(start, end, line_crs)
}
\arguments{
\item{start}{geos_geoemtry, point}

\item{end}{geos_geoemtry, point}

\item{line_crs}{crs}
}
\value{
geos_geometry linestring
}
\description{
Given 2 geos_geometry point geometries, create a line between the 2 points
}
