% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_progress_bar}
\alias{make_progress_bar}
\title{Make a progress bar and return an "make_progress()" function to update the progress bar.
Credit to the exactextractr team: https://github.com/isciences/exactextractr/blob/5fd17dcf02717332b125345aea586304f668cf12/R/exact_extract_helpers.R#L361}
\usage{
make_progress_bar(progress, n)
}
\arguments{
\item{progress}{logical, whether to make a progress bar or not (FALSE)}

\item{n}{numeric, total number of iterations}
}
\value{
make_progress function, when called will increment the progress bar text
}
\description{
Make a progress bar and return an "make_progress()" function to update the progress bar.
Credit to the exactextractr team: https://github.com/isciences/exactextractr/blob/5fd17dcf02717332b125345aea586304f668cf12/R/exact_extract_helpers.R#L361
}
\examples{
progress=TRUE
x = 1:500000
make_progress <- make_progress_bar(progress, length(x))
for (i in 1:length(x)) {
    make_progress()
}
}
