% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_geometry_to_last}
\alias{move_geometry_to_last}
\title{Move Geometry Column to the last column position}
\usage{
move_geometry_to_last(df)
}
\arguments{
\item{df}{A dataframe or an sf dataframe.}
}
\value{
Returns the input dataframe with the geometry column moved to the last position if it exists. Otherwise, returns the input dataframe as is.
}
\description{
Internal utility function for taking a dataframe or an sf dataframe, checks for the existence of a geometry type column, and
if it exists, moves it to the last column. If no geometry column exists, it returns the input dataframe as is.
}
\examples{
\dontrun{
# Create a dataframe
df <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
# Add a geometry column (sf dataframe)
df_sf <- sf::st_sf(df, geometry = sf::st_sfc(sf::st_point(c(1, 2, 3))))
# move column 
df_sf <- dplyr::relocate(df_sf, x, geometry, y)
df_sf # geometry column should be move to the middle column
# Move geometry column to last position
df_sf_moved <- move_geometry_to_last(df_sf)
df_sf_moved # geometry column should be move the end column
}
}
