% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_improvements.R
\name{needs_rectification}
\alias{needs_rectification}
\title{Check if there transects without valid banks or relief}
\usage{
needs_rectification(transects)
}
\arguments{
\item{transects}{sf linestring with "valid_banks" and "has_relief" logical columns}
}
\value{
logical, TRUE if there are transects without valid banks or relief
}
\description{
Check if there transects without valid banks or relief
}
