% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nextgen_braided_flowlines}
\alias{nextgen_braided_flowlines}
\title{NextGen Braided Flowlines}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 183 rows and 11 columns.
}
\source{
Created using hydrofabric3D.
}
\usage{
nextgen_braided_flowlines
}
\description{
A dataset of braided flowlines compatible with the NextGen hydrologic prediction system.
}
\keyword{datasets}
