% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_to_polygon_extensions.R
\name{pick_extension_pts}
\alias{pick_extension_pts}
\title{Decide the start and end points for the final transect line given two extended versions of the same transect
Requires two logicals indicating what to do with the extensions (these are decided by checking for intersections with the rest of the network)
Internal helper function}
\usage{
pick_extension_pts(left_extension, right_extension, use_left, use_right)
}
\arguments{
\item{left_extension}{geos_geometry linestring}

\item{right_extension}{geos_geometry linestring}

\item{use_left}{logical, do we use the left extension}

\item{use_right}{logical, do we use the right extension}
}
\value{
geos_geometry points, the start and end point of the final extension line
}
\description{
Decide the start and end points for the final transect line given two extended versions of the same transect
Requires two logicals indicating what to do with the extensions (these are decided by checking for intersections with the rest of the network)
Internal helper function
}
