% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_visualization.R
\name{plot_cs_pts}
\alias{plot_cs_pts}
\title{Plots an X-Y scatter plot of cross section points}
\usage{
plot_cs_pts(cs_pts, x = "pt_id", y = "Z", color = NULL, grid = FALSE)
}
\arguments{
\item{cs_pts}{data.frame of cross section points with columns hy_id, cs_id and columns for X and Y axises (i.e. "pt_id", "Z")}

\item{x}{character name of column in cs_pts to use for X axis}

\item{y}{character name of column in cs_pts to use for Y axis}

\item{color}{character name of column in cs_pts to color points on plot}

\item{grid}{logical, if TRUE then use facet_grid, otherwise use facet_wrap. Default is FALSE (uses facet_wrap)}
}
\value{
ggplot2 object
}
\description{
Plots an X-Y scatter plot of cross section points
}
