% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{prep_flowlines}
\alias{prep_flowlines}
\title{Prepare flowlines have a more dense and/or smoother surface for cutting transects}
\usage{
prep_flowlines(flowlines, densify = NULL, smooth = FALSE, verbose = TRUE)
}
\arguments{
\item{flowlines}{sf dataframe of flowline linestrings}

\item{densify}{numeric, if NULL, no densification happens. Default is NULL}

\item{smooth}{logical, whether to smooth linestrings}

\item{verbose}{logical}
}
\value{
sf dataframe
}
\description{
Prepare flowlines have a more dense and/or smoother surface for cutting transects
}
