% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{pts_to_reevaluate}
\alias{pts_to_reevaluate}
\title{Get a dataframe of points that should be evaluated due to their proximity (nearness in Z distance) to the bottom}
\usage{
pts_to_reevaluate(cs_pts, threshold = 1, pct_threshold = 0.99)
}
\arguments{
\item{cs_pts}{dataframe/sf dataframe of cross section points (requires hy_id, cs_id, and Z values)}

\item{threshold}{numeric, threshold distance in meters for points to be considered "near the bottom". Default is 1 meter (i.e. check if points are within 1 meter above the bottom)}

\item{pct_threshold}{numeric, threshold percent of points in the cross section that are within threshold of bottom to
determine whether point should be considered for re evaluation. Default is 0.99 (i.e. 99\% of points are near the bottom). Default is 0.99 (i.e. 99&\%).}
}
\value{
dataframe with the hy_id, cs_id, pts_near_bottom (count of pts_near_bottom), and pct_near_bottom (\% of points in cross section that are near bottom).
An empty dataframe is returned if ZERO points are classified as "near the bottom"
}
\description{
Get a dataframe of points that should be evaluated due to their proximity (nearness in Z distance) to the bottom
}
