% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pts_to_XY}
\alias{pts_to_XY}
\title{Convert an sf dataframe with a point geometry column to non spatial with XY columns}
\usage{
pts_to_XY(pts)
}
\arguments{
\item{pts}{sf dataframe of points}
}
\value{
data.frame or tibble with added X and Y columns
}
\description{
Convert an sf dataframe with a point geometry column to non spatial with XY columns
}
