% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{rectify_flat_cs_v1}
\alias{rectify_flat_cs_v1}
\title{Check for flat cross sections and try to update these values by extending the original cross sections and reextracting DEM values
(Deprecated version 1)}
\usage{
rectify_flat_cs_v1(
  cs_pts = NULL,
  net = NULL,
  cs = NULL,
  points_per_cs = NULL,
  min_pts_per_cs = 10,
  dem =
    "/vsicurl/https://prd-tnm.s3.amazonaws.com/StagedProducts/Elevation/13/TIFF/USGS_Seamless_DEM_13.vrt",
  scale = 0.5,
  threshold = 0
)
}
\arguments{
\item{cs_pts}{Output from extract_dem_values_first}

\item{net}{Hydrographic LINESTRING Network}

\item{cs}{character, Hydrographic LINESTRING Network file path}

\item{points_per_cs}{the desired number of points per CS. If NULL, then approximently 1 per grid cell resultion of DEM is selected.}

\item{min_pts_per_cs}{Minimun number of points per cross section required.}

\item{dem}{the DEM to extract data from}

\item{scale}{numeric, If a transect line DEM extraction results in all equal Z values,
by what percent of the transect lines length (meters) should the transect line be
extended in both directions to try to capture representative Z values ? Default is 0.5 (50\% of the transect length)}

\item{threshold}{numeric, threshold Z value (meters) that determines if a cross section is flat.
A threshold = 0 means if all Z values are the same, then the cross section is considered flat.
A threshold value of 1 means that any cross section with Z values all within 1 meter of eachother, is considered flat. Default is 0.}
}
\value{
sf object of cs_pts with "flat" cross sections removed/updated with longer transects to capture more Z data
}
\description{
Check for flat cross sections and try to update these values by extending the original cross sections and reextracting DEM values
(Deprecated version 1)
}
