% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_rectify.R
\name{rectify_summary}
\alias{rectify_summary}
\title{Makes a summaru dataframe and print out of differences between 2 cross section points dataframes}
\usage{
rectify_summary(input_points, output_points, verbose = TRUE)
}
\arguments{
\item{input_points}{sf dataframe or dataframe of cross section points}

\item{output_points}{sf dataframe or dataframe of cross section points, with "is_extended" logical column}

\item{verbose}{logical, whether to print out summary message/ Default is TRUE}
}
\value{
dataframe
}
\description{
Convenience function for printing out the difference between a cross section point dataframe and
the resulting output of putting that dataframe through the rectify_cs() function
}
