% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_cols_from_df}
\alias{remove_cols_from_df}
\title{Remove specified columns from a dataframe if they exist.}
\usage{
remove_cols_from_df(df, columns_to_remove)
}
\arguments{
\item{df}{A dataframe.}

\item{columns_to_remove}{character vector specifying the names of columns to be removed.}
}
\value{
dataframe with specified columns removed if they exist.
}
\description{
Remove specified columns from a dataframe if they exist.
}
