% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_rectify.R
\name{renumber_cs_ids}
\alias{renumber_cs_ids}
\title{Fix IDs in a dataframe}
\usage{
renumber_cs_ids(df)
}
\arguments{
\item{df}{A dataframe containing hy_id and cs_id columns.}
}
\value{
The input dataframe with renumbered cs_id values.
}
\description{
This function renumbers cross section IDs in a dataframe to ensure each hy_id has cross sections
numbered from 1 to the total number of cross sections on the hy_id.
}
