% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_rectify.R
\name{renumber_cs_ids2}
\alias{renumber_cs_ids2}
\title{Fix IDs in a dataframe}
\usage{
renumber_cs_ids2(df, crosswalk_id = NULL)
}
\arguments{
\item{df}{crosswalk_id character, name of primary ID column}
}
\value{
The input dataframe with renumbered cs_id values.
}
\description{
This function renumbers cross section IDs in a dataframe to ensure each crosswalk_id has cross sections
numbered from 1 to the total number of cross sections on the crosswalk_id.
}
