% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{rm_multi_intersects}
\alias{rm_multi_intersects}
\title{Selectively removes intersecting transect lines
Attempts to remove transects intersecting other transects by first removing transects that interesect the most other transects, then re checking intersection condition,  and doing this until there are no multi intersections
this gives the benefit of removing a transect line that intersects many other transects, potentially leaving those other transects with no extraneous intersections ONCE the MULTI intersecting transect is removed}
\usage{
rm_multi_intersects(x)
}
\arguments{
\item{x}{sf dataframe of linestrings}
}
\value{
sf dataframe
}
\description{
Selectively removes intersecting transect lines
Attempts to remove transects intersecting other transects by first removing transects that interesect the most other transects, then re checking intersection condition,  and doing this until there are no multi intersections
this gives the benefit of removing a transect line that intersects many other transects, potentially leaving those other transects with no extraneous intersections ONCE the MULTI intersecting transect is removed
}
