% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{rm_multiflowline_intersections}
\alias{rm_multiflowline_intersections}
\title{Remove transect lines that intersect with more than one flowline}
\usage{
rm_multiflowline_intersections(transects, flowlines)
}
\arguments{
\item{transects}{sf linestring dataframe of transect lines}

\item{flowlines}{sf linestring dataframe of flowlines}
}
\value{
sf linestring dataframe
}
\description{
Remove transect lines that intersect with more than one flowline
}
