% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rm_self_intersections}
\alias{rm_self_intersections}
\title{Remove Self-Intersections}
\format{
An object of class \code{function} of length 1.
}
\source{
Generated using \code{rm_self_intersections} function.
}
\usage{
rm_self_intersections
}
\description{
A dataset of flowlines processed to remove self-intersections.
}
\keyword{datasets}
