% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_cs_pts}
\alias{select_cs_pts}
\title{Select standard cross section point columns
Internal helper function for selecting cross section point columns aligning with standard data model for cross section points}
\usage{
select_cs_pts(cs_pts, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts}{dataframe, tibble, or sf dataframe}

\item{crosswalk_id}{character, unique ID column}
}
\value{
dataframe, tibble, or sf dataframe with only relavent cross section point columns
}
\description{
Select standard cross section point columns
Internal helper function for selecting cross section point columns aligning with standard data model for cross section points
}
