% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_transects}
\alias{select_transects}
\title{Select standard transect columns
Internal helper function for selecting transect columns aligning with standard data model for transect}
\usage{
select_transects(transects, crosswalk_id = NULL)
}
\arguments{
\item{transects}{dataframe, tibble, or sf dataframe}

\item{crosswalk_id}{character, unique ID column}
}
\value{
dataframe, tibble, or sf dataframe with only relavent transects columns
}
\description{
Select standard transect columns
Internal helper function for selecting transect columns aligning with standard data model for transect
}
