% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{set_bank_anchors}
\alias{set_bank_anchors}
\title{Given a vector of depths and vector of corresponding point types representing a cross section of depths,}
\usage{
set_bank_anchors(depths, point_types, L, R)
}
\arguments{
\item{depths}{numeric, depth values}

\item{point_types}{character, poitn type}

\item{L}{numeric, left pointer}

\item{R}{numeric, right pointer}
}
\value{
character vector
}
\description{
Conditions for setting left and right bank classes:
\itemize{
\item make sure there is a left_bank at the left most highest position of the left side
\item make sure there is a "right_bank" at the right most highest position of the right side
}
}
