% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{set_channel_anchors}
\alias{set_channel_anchors}
\title{Set a "channel" point on the left and right ends of the "bottom" in a set of cross section point types
Make sure that the "bottom" has ATLEAST 1 channel point on both the left and right sides of the bottom}
\usage{
set_channel_anchors(point_types)
}
\arguments{
\item{point_types}{character vector}
}
\value{
character vector
}
\description{
Set a "channel" point on the left and right ends of the "bottom" in a set of cross section point types
Make sure that the "bottom" has ATLEAST 1 channel point on both the left and right sides of the bottom
}
