% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{set_channel_surrounded_by_bottom}
\alias{set_channel_surrounded_by_bottom}
\title{Set any channel points that are entirely surrounded by bottom points to bottom points}
\usage{
set_channel_surrounded_by_bottom(depths, point_types)
}
\arguments{
\item{depths}{numeric vector}

\item{point_types}{character vector}
}
\value{
character
}
\description{
Set any channel points that are entirely surrounded by bottom points to bottom points
}
