% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{set_left_bank}
\alias{set_left_bank}
\title{Set any points to the left of the right most left bank point to "left_bank" IF there is a channel then a bottom to the right of the left_bank point}
\usage{
set_left_bank(point_types)
}
\arguments{
\item{point_types}{character vector}
}
\value{
character
}
\description{
Set any points to the left of the right most left bank point to "left_bank" IF there is a channel then a bottom to the right of the left_bank point
}
