% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{set_missing_bottom}
\alias{set_missing_bottom}
\title{Set an artificial bottom at the lowest channel point if cross section points are in a left_bank -> channel -> right_bank pattern
Only applies to cross section points where there is NO bottom and
the point types are in this order: left_bank -> channel -> right_bank}
\usage{
set_missing_bottom(depths, point_types)
}
\arguments{
\item{depths}{numeric vector}

\item{point_types}{character vector}
}
\value{
character vector
}
\description{
Set an artificial bottom at the lowest channel point if cross section points are in a left_bank -> channel -> right_bank pattern
Only applies to cross section points where there is NO bottom and
the point types are in this order: left_bank -> channel -> right_bank
}
