% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{set_right_bank}
\alias{set_right_bank}
\title{Set any points to the right of the left most right bank point to "right_bank" IF there is a channel then a bottom to the left of the right_bank point}
\usage{
set_right_bank(point_types)
}
\arguments{
\item{point_types}{character vector}
}
\value{
character
}
\description{
Set any points to the right of the left most right bank point to "right_bank" IF there is a channel then a bottom to the left of the right_bank point
}
