% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_without_network.R
\name{shorten_flagged_transects}
\alias{shorten_flagged_transects}
\title{Shorten specific flagged transects by specified distance
Shorten transects by 'extension_distance' if they have a 'flagged' column value of TRUE}
\usage{
shorten_flagged_transects(transects, crosswalk_id = NULL)
}
\arguments{
\item{transects}{sf dataframe of LINESTRINGS, requires an 'extension_distance' column to specify how much to flagged transects by}

\item{crosswalk_id}{character, unique ID column}
}
\value{
sf dataframe of transects with shortened transects where flagged is TRUE
}
\description{
Shorten specific flagged transects by specified distance
Shorten transects by 'extension_distance' if they have a 'flagged' column value of TRUE
}
