% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_without_network.R
\name{shorten_multi_transect_intersecting_extended_transects}
\alias{shorten_multi_transect_intersecting_extended_transects}
\title{Takes any transects with multiple intersections that was extended, and shortens them by the distance specified in the "extension_distance" column}
\usage{
shorten_multi_transect_intersecting_extended_transects(x, crosswalk_id = NULL)
}
\arguments{
\item{x}{sf dataframe of transects, requires a crosswalk_id, cs_id, cs_lengthm, extension_distance, and geometry column}

\item{crosswalk_id}{character, unique ID column}
}
\value{
sf dataframe of transects with any transects that intersect multiple other transects being shortened by -extension_distance
}
\description{
Takes any transects with multiple intersections that was extended, and shortens them by the distance specified in the "extension_distance" column
}
