% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{smooth_depths}
\alias{smooth_depths}
\title{Smooth out depth values}
\usage{
smooth_depths(depths, num_of_pts = NULL, window = 3)
}
\arguments{
\item{depths}{}

\item{num_of_pts}{}

\item{window}{}
}
\value{
numeric vector
}
\description{
Smooth out depth values
}
