% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_utils.R
\name{subset_polygons_in_transects}
\alias{subset_polygons_in_transects}
\title{Get polygons that intersect with the transects}
\usage{
subset_polygons_in_transects(transect_lines, polygons)
}
\arguments{
\item{transect_lines}{Set of Sf linestrigns to extend (only if the transect lines are ENTIRELLY within a polygons)}

\item{polygons}{set of sf polygons that transect lines should be exteneded}
}
\value{
sf polygon dataframe
}
\description{
Get polygons that intersect with the transects
}
