% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{transects_to_cs_pts}
\alias{transects_to_cs_pts}
\title{Convert SF linestring transect lines into SF points with}
\usage{
transects_to_cs_pts(transects, points_per_cs)
}
\arguments{
\item{transects}{sf linestring}

\item{points_per_cs}{numeric vector of length 'transects', indicating the number of points to get per transect}
}
\value{
sf point dataframe
}
\description{
Convert SF linestring transect lines into SF points with
}
