% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_to_polygon_extender.R
\name{trim_transects_to_polygons}
\alias{trim_transects_to_polygons}
\title{Trim a set of transects to the bounds of polygons}
\usage{
trim_transects_to_polygons(
  transect_lines,
  flowlines,
  polygons,
  crosswalk_id = NULL,
  dissolve = FALSE
)
}
\arguments{
\item{transect_lines}{sf dataframe}

\item{flowlines}{sf dataframe}

\item{polygons}{sf dataframe}

\item{crosswalk_id}{character unique ID}

\item{dissolve}{logical, whether to dissolve polygon internal boundaries or not. Default is FALSE.}
}
\value{
sf dataframe
}
\description{
Trim a set of transects to the bounds of polygons
}
