% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{unique_braids}
\alias{unique_braids}
\title{Create a list of braid IDs containing COMIDs in each braid for a single continguous network}
\usage{
unique_braids(network, start = NULL, verbose = FALSE)
}
\arguments{
\item{network}{The network object representing the river network.}

\item{start}{Optional argument specifying the starting point for braid detection.}

\item{verbose}{Logical indicating whether to display verbose messages during the braid detection process.}
}
\value{
list of braid IDs with COMIDs within each braid
}
\description{
Find and uniquely identify braids in a network of flowlines, given a dataframe containing comid, fromnode, tonode and divergence as columns. 'find_braids()" identifies braids as cycles in the graph representation of the river network.
Internal function for use in 'get_braid_list()'
}
