% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_pts.R
\name{use_smoothed_depths}
\alias{use_smoothed_depths}
\title{Use the bottom points from the original depth values otherwise use the (rolling mean) smoothed depth values}
\usage{
use_smoothed_depths(start_depths, smoothed_depths, point_types)
}
\arguments{
\item{start_depths}{}

\item{smoothed_depths}{}

\item{point_types}{}
}
\value{
numeric vector
}
\description{
Use the bottom points from the original depth values otherwise use the (rolling mean) smoothed depth values
}
