% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_arg_types}
\alias{validate_arg_types}
\title{Validate all arguments given are of the types specified in type_map
Function throws an error if any argument is not specified in the type map or its type does not match the possible types in type_map
Internal function for easily type checking all the inputs for a given function}
\usage{
validate_arg_types(..., type_map)
}
\arguments{
\item{...}{list of args}

\item{type_map}{list, argument names as the keys and a vector of values representing the valid types}
}
\value{
logical, TRUE if no error, otherwise raises an error
}
\description{
Validate all arguments given are of the types specified in type_map
Function throws an error if any argument is not specified in the type map or its type does not match the possible types in type_map
Internal function for easily type checking all the inputs for a given function
}
