% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_classified_cs_pts}
\alias{validate_classified_cs_pts}
\title{Validate Classified Cross Sections Points
Ensure all cross section points are valid. This validates the points in the same manner as validate_cs_pts() but also checks that
classification columns ('class', 'point_type', 'valid_banks', 'has_relief') exist.}
\usage{
validate_classified_cs_pts(cs_pts, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts}{sf object, cross section points}

\item{crosswalk_id}{character, column name of the crosswalk id}
}
\value{
logical, TRUE if cs_pts meet all required criteria, FALSE otherwise
}
\description{
Validate Classified Cross Sections Points
Ensure all cross section points are valid. This validates the points in the same manner as validate_cs_pts() but also checks that
classification columns ('class', 'point_type', 'valid_banks', 'has_relief') exist.
}
