% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_cs_pts}
\alias{validate_cs_pts}
\title{Validate Cross Sections Points
Ensure all cross section points are valid}
\usage{
validate_cs_pts(cs_pts, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts}{sf object, cross section points}

\item{crosswalk_id}{character, column name of the crosswalk id}
}
\value{
logical, TRUE if cs_pts meet all required criteria, FALSE otherwise
}
\description{
Validate Cross Sections Points
Ensure all cross section points are valid
}
