% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_cut_cross_section_inputs}
\alias{validate_cut_cross_section_inputs}
\title{Validate Inputs for cut_cross_sections Function}
\usage{
validate_cut_cross_section_inputs(
  net,
  id,
  cs_widths,
  num,
  smooth,
  densify,
  rm_self_intersect,
  fix_braids,
  braid_threshold,
  braid_method,
  precision,
  add
)
}
\arguments{
\item{net}{An sf object representing the hydrographic network.}

\item{id}{A unique identifier column in the network data.}

\item{cs_widths}{Bankfull widths (length of cross sections) for each network element.}

\item{num}{Number of transects per network element.}

\item{smooth}{Logical, whether to smooth linestring geometries or not.}

\item{densify}{Numeric, the factor by which to densify the linestrings.}

\item{rm_self_intersect}{Logical, whether to remove self-intersecting transect linestrings.}

\item{fix_braids}{Logical, whether to fix braided transect lines or not.}

\item{braid_threshold}{Numeric, the total length of all flowlines in a braid below which fix_braid_transects should operate.}

\item{braid_method}{Character, the method to determine the geometries to cut. Options are "comid", "component", or "neighbor". Default is "comid".}

\item{precision}{Numeric, the number of meters to approximate final cross-section linestring length.}

\item{add}{Logical, indicating whether to add original 'net' data to the outputted transect lines.}
}
\value{
NULL if inputs are valid; otherwise, an error is thrown.
}
\description{
This function validates the inputs for the cut_cross_sections function to ensure they meet the required criteria.
}
\keyword{internal}
