% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{validate_df}
\alias{validate_df}
\title{Validate that a dataframe is valid type (inherits from a dataframe) and has the given required columns}
\usage{
validate_df(x, cols, obj_name = NULL)
}
\arguments{
\item{x}{dataframe, sf, or tibble}

\item{cols}{character vector of columns that are required in 'x'}

\item{obj_name}{character string, optional name of object to display in error messages. If not obj_name is given, obj_name defaults to 'x'}
}
\value{
logical, TRUE value if all conditions are met and valid, otherwise an error is raised
}
\description{
Validate that a dataframe is valid type (inherits from a dataframe) and has the given required columns
}
