% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_transects}
\alias{validate_transects}
\title{Validate Transects}
\usage{
validate_transects(transects, crosswalk_id = NULL)
}
\arguments{
\item{transects}{sf object, transects}

\item{crosswalk_id}{character, column name of the crosswalk id}
}
\value{
logical, TRUE if all validations pass, FALSE otherwise
}
\description{
Validate Transects
}
