% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_transects_against_flowlines}
\alias{validate_transects_against_flowlines}
\title{Validate Transects Against Flowlines
Ensure all transects are valid relative to a set of flowlines}
\usage{
validate_transects_against_flowlines(transects, flowlines, crosswalk_id = NULL)
}
\arguments{
\item{transects}{sf object, transects}

\item{flowlines}{sf object, flowlines}

\item{crosswalk_id}{character, column name of the crosswalk id}
}
\value{
logical, TRUE if all validations pass, FALSE otherwise
}
\description{
Validate Transects Against Flowlines
Ensure all transects are valid relative to a set of flowlines
}
