% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_bathymetry.R
\name{add_cs_bathymetry}
\alias{add_cs_bathymetry}
\title{Given provide inchannel widths and depths to a set of cross section points and derive estimated shapes}
\usage{
add_cs_bathymetry(cs_pts = NULL, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts}{dataframe or sf dataframe. Default is NULL}

\item{crosswalk_id}{character, ID column}
}
\value{
dataframe or sf dataframe with AHG estimated points injected into the input cross section points
}
\description{
Takes in a point of cross section points with added top width (TW), depth (DEPTH), and dingman_r (DINGMAN_R) columns
}
