% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{cut_cross_sections}
\alias{cut_cross_sections}
\title{Generate Cross Sections Across Hydrographic Network}
\usage{
cut_cross_sections(
  net,
  id = NULL,
  cs_widths = 100,
  num = 10,
  smooth = TRUE,
  densify = 2,
  rm_self_intersect = TRUE,
  fix_braids = TRUE,
  terminal_id = NULL,
  braid_threshold = NULL,
  version = 2,
  braid_method = "comid",
  add = FALSE
)
}
\arguments{
\item{net}{Hydrographic LINESTRING Network}

\item{id}{Unique Identifier in net}

\item{cs_widths}{Bankfull Widths (length of cross sections for each net element)}

\item{num}{Number of transects per Net element}

\item{smooth}{logical, whether to smooth linestring geometries or not. Default is TRUE.}

\item{densify}{numeric, how many times more points should be added to linestrings. Default is 2.}

\item{rm_self_intersect}{logical, whether to remove self intersecting transect linestrings}

\item{fix_braids}{logical, whether to fix braided transect lines or not. If TRUE (default), linestrings that are part of a braided network are augmented}

\item{terminal_id}{character, column name containing a unique identifier, delineating separate networks in the 'net' dataset. Default is NULL which will best effort determine the connected components in the network to try and create a 'component_id' column in 'net'}

\item{braid_threshold}{numeric value, value of the total length of all flowlines in a braid. Only braids with total flowline
lengths less than or equal to the threshold will be considered by function (i.e. determines that maximum braid size that fix_braid_transects() should operate on).
Default is NULL, which will attempt to fix all the braid transects in the data}

\item{version}{integer, version number of braid algorithm to use, either 1 or 2. Default is 2.}

\item{braid_method}{The method to determine the geometries to cut. Options are "comid", "component", or "neighbor". Default is "comid"}

\item{add}{logical indicating whether to add original 'net' data to the outputted transect lines. Default is FALSE.}
}
\value{
sf object
}
\description{
Generate Cross Sections Across Hydrographic Network
}
