% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_without_network.R
\name{extend_by_percent}
\alias{extend_by_percent}
\title{Extend an sf linestring dataframe by a percent of the lines length}
\usage{
extend_by_percent(x, crosswalk_id = NULL, pct = 0.5, length_col = NULL)
}
\arguments{
\item{x}{linestring sf dataframe}

\item{crosswalk_id}{character, unique ID column name}

\item{pct}{numeric, percent of line to extend linestring by in both directions}

\item{length_col}{character, name of the output length column name. Default is NULL which will create a length column name of "geom_length".}
}
\value{
sf dataframe with extended linestring geometries
}
\description{
Extend an sf linestring dataframe by a percent of the lines length
}
