% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_extenders_without_network.R
\name{extend_transects}
\alias{extend_transects}
\title{Extend a set of transects by a percentage}
\usage{
extend_transects(
  transects_to_extend,
  transects,
  net,
  crosswalk_id = NULL,
  scale = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{transects_to_extend}{sf linestrings, set of transects that should be extended (subset of 'transects'). Requires the following columns:  "hy_id", "cs_id", "cs_lengthm" (length of geometry in meters)}

\item{transects}{sf linestrings, set of all transects in the network. Requires the following columns: "hy_id", "cs_id", "cs_lengthm" (length of geometry in meters)}

\item{net}{sf linestrings, flowline network that transects were generated from, requires "id" column (where "id" equals the "hy_id" columns in 'transects_to_extend' and 'transects' )}

\item{crosswalk_id}{character, unique ID column name}

\item{scale}{numeric, percentage of current transect line length to extend transects in transects_to_extend by. Default is 0.5 (50\% of the transect length)}

\item{verbose}{logical, whether to print messages or not. Default is TRUE}
}
\value{
sf linestring dataframe containing the updates transects_to_extend (with a flag denoting if the geometry was extended by "scale" \% or not)
}
\description{
Extend a set of transects by a percentage
}
