% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_improvements.R
\name{extend_transects_by_cs_attributes}
\alias{extend_transects_by_cs_attributes}
\title{Extend transects for any transects with invalid cross section attributes}
\usage{
extend_transects_by_cs_attributes(
  flowlines = NULL,
  transects = NULL,
  crosswalk_id = NULL,
  scale = 0.5,
  keep_lengths = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{flowlines}{dataframe or sf dataframe}

\item{transects}{dataframe or sf dataframe}

\item{crosswalk_id}{character}

\item{scale}{numeric}

\item{keep_lengths}{logical whether to keep a record of the original transect lengths or not, default is FALSE, original lengths are not kept}

\item{verbose}{logical}
}
\value{
dataframe or sf dataframe
}
\description{
Extend transects for any transects with invalid cross section attributes
}
