% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{find_braids}
\alias{find_braids}
\title{Find braided sections of a network and return the unique crosswalk_ids for each idenfied braid}
\usage{
find_braids(network, crosswalk_id = NULL, nested = TRUE, verbose = FALSE)
}
\arguments{
\item{network}{The network object representing the river network.}

\item{crosswalk_id}{unique ID column name}

\item{nested}{Logical indicating whether the output dataframe should be nested, with each COMID having a list of all the braids it is a part of. If TRUE (Default), the braid_id column may contain multiple braid IDs for a given COMID. If FALSE, there may be duplicate COMIDs as a single COMID could be a part of multiple braids (braid_id)}

\item{verbose}{Logical indicating whether to display verbose messages during the braid detection process.}
}
\value{
dataframe or sf dataframe with added braid_id
}
\description{
Find and uniquely identify braids in a network of flowlines, given a dataframe containing comid, fromnode, tonode and divergence as columns. 'find_braids()" identifies braids as cycles in the graph representation of the river network.
}
\examples{
 \dontrun{
net <- nhdplusTools::navigate_network(
 start       = 101,
 mode        = "UT",
 distance_km = 100
 ) 
 
# drop most of the columns in the network dataset
net <- dplyr::select(net, comid, divergence, totdasqkm, fromnode, tonode, terminalpa)

# get a dataframe of COMIDs and braid IDs
braids <- find_braids(network = net, add = FALSE)

# add braid_id column to original dataset
braid_df = find_braids(network   = net,
                       nested    = TRUE,
                       )

# returns original data with each braid_id represented
# by its individual COMIDs (may contain duplicate COMIDs)
nested_braids = find_braids(network   = net,
                       nested    = FALSE
                       )
}
}
