% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_transects.R
\name{fix_braid_transects2}
\alias{fix_braid_transects2}
\title{Fix transects found on braided river sections}
\usage{
fix_braid_transects2(
  net,
  transect_lines,
  terminal_id = NULL,
  braid_threshold = NULL
)
}
\arguments{
\item{net}{sf object of NHDplusv2 data}

\item{transect_lines}{sf linestring dataframe, containing cross sections of flowlines in 'net' the output of "cut_cross_sections2()" function}

\item{terminal_id}{character, column name containing a unique identifier, delineating seperate networks in the 'network' dataset. Default is NULL which will use 'find_connected_components()' and determine the connected components in the graph to try and create a 'component_id' column in 'network'}

\item{braid_threshold}{numeric value, value of the total length of all flowlines in a braid. Only braids with total flowline
lengths less than or equal to the threshold will be considered by function (i.e. determines that maximum braid size that fix_braid_transects() should operate on).
Default is NULL, which will attempt to fix all the braid transects in the data}
}
\value{
sf object of transect linestrings
}
\description{
Fix transects found on braided river sections
}
