% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_improvements.R
\name{flag_transects_for_change}
\alias{flag_transects_for_change}
\title{Add a flagged and extension distance columns to set of transects with CS attributes based on new cross section points data}
\usage{
flag_transects_for_change(
  x,
  crosswalk_id = NULL,
  points_per_cs = NULL,
  min_pts_per_cs = 10,
  dem = default_dem,
  pct_of_length_for_relief = 0.01,
  na.rm = TRUE
)
}
\arguments{
\item{x}{sf dataframe of transects}

\item{crosswalk_id}{character, unique ID column}

\item{points_per_cs}{numeric}

\item{min_pts_per_cs}{numeric}

\item{dem}{character}

\item{pct_of_length_for_relief}{numeric}

\item{na.rm}{logical, whether to remove NAs from the given cross section points and any NA comparison points pulled from the dem. Default is TRUE}
}
\value{
sf dataframe of transects with updated geometries
}
\description{
Add a flagged and extension distance columns to set of transects with CS attributes based on new cross section points data
}
