% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braids.R
\name{get_braid_list}
\alias{get_braid_list}
\title{Create a list of braid IDs containing crosswalk_ids in each braid}
\usage{
get_braid_list(network, crosswalk_id = NULL, verbose = FALSE)
}
\arguments{
\item{network}{The network object representing the river network.}

\item{crosswalk_id}{unique ID column name}

\item{verbose}{Logical indicating whether to display verbose messages during the braid detection process.}
}
\value{
list of braid IDs and COMIDs within each braid
}
\description{
Find and uniquely identify braids in a network of flowlines, given an sf dataframe containing crosswalk_id and sf linestring geometries, 'find_braids()" identifies braids as cycles in the graph representation of the river network.
}
\examples{
 \dontrun{
net <- nhdplusTools::navigate_network(
 start       = 101,
 mode        = "UT",
 distance_km = 100
 ) 

net <- dplyr::select(net, comid, divergence, totdasqkm, fromnode, tonode, terminalpa)

# get a dataframe of COMIDs and braid IDs
braids <- get_braid_list(network = net, crosswalk_id = "comid")
}
}
