% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_bathymetry.R
\name{get_cs_bottom_length}
\alias{get_cs_bottom_length}
\title{Calculate the length between the leftmost and rightmost bottom point in each cross section}
\usage{
get_cs_bottom_length(cs_pts, crosswalk_id = NULL)
}
\arguments{
\item{cs_pts}{dataframe, or sf dataframe of cross section points}

\item{crosswalk_id}{character, ID column}
}
\value{
summarized dataframe of input cs_pts dataframe with a bottom_length value for each hy_id/cs_id
}
\description{
Calculate the length between the leftmost and rightmost bottom point in each cross section
}
