% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_utils.R
\name{get_transect_extension_distances_to_polygons}
\alias{get_transect_extension_distances_to_polygons}
\title{Give a set of transecct linestrings and poylgons and get the minimum distance to extend each transect line (from both directions, to try and reach the edge of a "polygons")
WIP/internal function for extending transect lines out to FEMA 100 year flood plain polygons
DEPRECATED at this point, will delete on next version}
\usage{
get_transect_extension_distances_to_polygons(
  transect_lines,
  polygons,
  flines,
  max_extension_distance
)
}
\arguments{
\item{transect_lines}{Set of Sf linestrigns to extend (only if the transect lines are ENTIRELLY within a polygons)}

\item{polygons}{set of sf polygons that transect lines should be exteneded}

\item{flines}{set of Sf linestrings}

\item{max_extension_distance}{numeric, maximum distance (meters) to extend a transect line in either direction to try and intersect one of the "polygons"}
}
\value{
sf linestring, with extended transect lines
}
\description{
Give a set of transecct linestrings and poylgons and get the minimum distance to extend each transect line (from both directions, to try and reach the edge of a "polygons")
WIP/internal function for extending transect lines out to FEMA 100 year flood plain polygons
DEPRECATED at this point, will delete on next version
}
