% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{get_transects}
\alias{get_transects}
\title{Generate Multiple cross section along a linestring}
\usage{
get_transects(edges, line, bf_width)
}
\arguments{
\item{edges}{data.frame of LINESTRINGs (pieces of line)}

\item{line}{original line element}

\item{bf_width}{Bankfull Width (length of cross section)}
}
\value{
GEOS object
}
\description{
Generate Multiple cross section along a linestring
}
