% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{linestring}
\alias{linestring}
\title{Tester}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 325 rows and 5 columns.
}
\usage{
linestring
}
\description{
Tester
}
\keyword{datasets}
